# combiner/server/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: The Place of Interest, Event, and Organism API
  description: Generic Poi, Event, and Organism API server
  version: "1.0"
servers:
- url: http://0.0.0.0:8081
paths:
...
  /things/events/deactivate/{path}:
    put:
      tags:
      - events
      summary: deactivate events
      description: Delete an Event
      operationId: events#deactivate
      parameters:
      - name: path
        in: path
        description: Path
        required: true
        schema:
          type: string
          description: Path
          example: /croydon/purley/59b287534defd02dace3671b
        example: /croydon/purley/59b287534defd02dace3671b
      responses:
        "200":
          description: OK response.
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
...
  /things/events/list/{path}:
    get:
      tags:
      - events
      summary: ListByPath events
      description: List Events for a path
      operationId: events#ListByPath
      parameters:
      - name: path
        in: path
        description: Path
        required: true
        schema:
          type: string
          description: Path
          example: /croydon/purley/59b287534defd02dace3671b
        example: /croydon/purley/59b287534defd02dace3671b
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                description: 'Response body may alternatively be #/components/schemas/CspaceEventResponseSummaryCollection'
                $ref: '#/components/schemas/CspaceEventCollection'
              example:
              - deactivated: false
                description: Unde ad dolores sit sit quod enim.
                end: "2017-08-22T12:06:30.696Z"
                id: 59b287534defd02dace3671b
                path: Quasi consequuntur.
                properties:
                  Asperiores culpa quia.: Aperiam ea minima autem consequatur ducimus
                    ut.
                refs:
                - Magni et facere sapiente voluptatem sunt quis.
                - Repudiandae voluptatem est.
                - Dolores consequatur ex repellendus.
                start: "2017-08-22T12:06:30.696Z"
                title: Veniam cumque aut.
              - deactivated: false
                description: Unde ad dolores sit sit quod enim.
                end: "2017-08-22T12:06:30.696Z"
                id: 59b287534defd02dace3671b
                path: Quasi consequuntur.
                properties:
                  Asperiores culpa quia.: Aperiam ea minima autem consequatur ducimus
                    ut.
                refs:
                - Magni et facere sapiente voluptatem sunt quis.
                - Repudiandae voluptatem est.
                - Dolores consequatur ex repellendus.
                start: "2017-08-22T12:06:30.696Z"
                title: Veniam cumque aut.
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
...
  /things/organisms/deactivate/{path}:
    put:
      tags:
      - organisms
      summary: deactivate organisms
      description: Delete an Organism
      operationId: organisms#deactivate
      parameters:
      - name: path
        in: path
        description: Path
        required: true
        schema:
          type: string
          description: Path
          example: /croydon/charity
        example: /croydon/charity
      responses:
        "200":
          description: OK response.
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
...
  /things/organisms/list/{path}:
    get:
      tags:
      - organisms
      summary: ListByPath organisms
      description: List Orgs for a Path
      operationId: organisms#ListByPath
      parameters:
      - name: path
        in: path
        description: the path of the Org
        required: true
        schema:
          type: string
          description: the path of the Org
          example: Sunt voluptas et sit nostrum et.
        example: Magnam aut voluptatem nisi.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                description: 'Response body may alternatively be #/components/schemas/CspaceOrganismResponseSummaryCollection'
                $ref: '#/components/schemas/CspaceOrganismCollection'
              example:
              - deactivated: true
                description: Natus ratione quia molestiae excepturi qui est.
                id: 59b287534defd02dace3671b
                path: Excepturi occaecati magni minus aut.
                properties:
                  Voluptate temporibus sapiente.: Sapiente dolores dicta.
                  Voluptatum aut non quaerat.: Quis quo ad qui ut corporis.
                refs:
                - Assumenda autem et nostrum dolorem.
                - Explicabo ut.
                - Quia cupiditate molestiae hic labore illo voluptas.
                title: Doloremque aut itaque iure.
              - deactivated: true
                description: Natus ratione quia molestiae excepturi qui est.
                id: 59b287534defd02dace3671b
                path: Excepturi occaecati magni minus aut.
                properties:
                  Voluptate temporibus sapiente.: Sapiente dolores dicta.
                  Voluptatum aut non quaerat.: Quis quo ad qui ut corporis.
                refs:
                - Assumenda autem et nostrum dolorem.
                - Explicabo ut.
                - Quia cupiditate molestiae hic labore illo voluptas.
                title: Doloremque aut itaque iure.
              - deactivated: true
                description: Natus ratione quia molestiae excepturi qui est.
                id: 59b287534defd02dace3671b
                path: Excepturi occaecati magni minus aut.
                properties:
                  Voluptate temporibus sapiente.: Sapiente dolores dicta.
                  Voluptatum aut non quaerat.: Quis quo ad qui ut corporis.
                refs:
                - Assumenda autem et nostrum dolorem.
                - Explicabo ut.
                - Quia cupiditate molestiae hic labore illo voluptas.
                title: Doloremque aut itaque iure.
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
...
  /things/organisms/refs/{path}:
    get:
      tags:
      - organisms
      summary: ListByReference organisms
      description: List Orgs for a Ref
      operationId: organisms#ListByReference
      parameters:
      - name: path
        in: path
        description: the path of the Org
        required: true
        schema:
          type: string
          description: the path of the Org
          example: Doloremque in ex laudantium culpa non.
        example: Doloremque error rerum quibusdam sint praesentium.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                description: 'Response body may alternatively be #/components/schemas/CspaceOrganismResponseSummaryCollection'
                $ref: '#/components/schemas/CspaceOrganismCollection'
              example:
              - deactivated: true
                description: Natus ratione quia molestiae excepturi qui est.
                id: 59b287534defd02dace3671b
                path: Excepturi occaecati magni minus aut.
                properties:
                  Voluptate temporibus sapiente.: Sapiente dolores dicta.
                  Voluptatum aut non quaerat.: Quis quo ad qui ut corporis.
                refs:
                - Assumenda autem et nostrum dolorem.
                - Explicabo ut.
                - Quia cupiditate molestiae hic labore illo voluptas.
                title: Doloremque aut itaque iure.
              - deactivated: true
                description: Natus ratione quia molestiae excepturi qui est.
                id: 59b287534defd02dace3671b
                path: Excepturi occaecati magni minus aut.
                properties:
                  Voluptate temporibus sapiente.: Sapiente dolores dicta.
                  Voluptatum aut non quaerat.: Quis quo ad qui ut corporis.
                refs:
                - Assumenda autem et nostrum dolorem.
                - Explicabo ut.
                - Quia cupiditate molestiae hic labore illo voluptas.
                title: Doloremque aut itaque iure.
              - deactivated: true
                description: Natus ratione quia molestiae excepturi qui est.
                id: 59b287534defd02dace3671b
                path: Excepturi occaecati magni minus aut.
                properties:
                  Voluptate temporibus sapiente.: Sapiente dolores dicta.
                  Voluptatum aut non quaerat.: Quis quo ad qui ut corporis.
                refs:
                - Assumenda autem et nostrum dolorem.
                - Explicabo ut.
                - Quia cupiditate molestiae hic labore illo voluptas.
                title: Doloremque aut itaque iure.
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
...
  /things/pois/deactivate/{path}:
    put:
      tags:
      - pois
      summary: deactivate pois
      description: Update a POI
      operationId: pois#deactivate
      parameters:
      - name: path
        in: path
        description: the path of the POI
        required: true
        schema:
          type: string
          description: the path of the POI
          example: In eveniet eum.
        example: Eveniet rerum quia id dolores excepturi.
      responses:
        "200":
          description: OK response.
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
...
  /things/pois/paths/{path}:
    get:
      tags:
      - pois
      summary: ListByPath pois
      description: List POIs for a path
      operationId: pois#ListByPath
      parameters:
      - name: path
        in: path
        description: the path of the POI
        required: true
        schema:
          type: string
          description: the path of the POI
          example: Laudantium nobis saepe voluptas et.
        example: Excepturi consequatur illum quas ut.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                description: 'Response body may alternatively be #/components/schemas/CspacePoiResponseSummaryCollection'
                $ref: '#/components/schemas/CspacePoiCollection'
              example:
              - deactivated: true
                description: Totam ut harum.
                geojson: Modi sunt dolorem explicabo id quia aut.
                id: 59b287534defd02dace3671b
                location: Similique sed rem quas et quo.
                path: Dolor dignissimos dolore.
                properties:
                  Et asperiores occaecati aut ut unde id.: Minima architecto officiis
                    et possimus ratione.
                  Porro quia eveniet similique autem.: Deleniti magnam veniam ad.
                  Sunt quia placeat.: Eveniet est facere.
                refs:
                - Vel soluta.
                - Qui qui omnis culpa quis atque sed.
                - Sed quia dolorem et.
                title: Magni aut et itaque iusto.
              - deactivated: true
                description: Totam ut harum.
                geojson: Modi sunt dolorem explicabo id quia aut.
                id: 59b287534defd02dace3671b
                location: Similique sed rem quas et quo.
                path: Dolor dignissimos dolore.
                properties:
                  Et asperiores occaecati aut ut unde id.: Minima architecto officiis
                    et possimus ratione.
                  Porro quia eveniet similique autem.: Deleniti magnam veniam ad.
                  Sunt quia placeat.: Eveniet est facere.
                refs:
                - Vel soluta.
                - Qui qui omnis culpa quis atque sed.
                - Sed quia dolorem et.
                title: Magni aut et itaque iusto.
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
...
components:
  schemas:
...
    CspaceEventCollection:
      type: array
      items:
        $ref: '#/components/schemas/CspaceEvent'
      example:
      - deactivated: false
        description: Unde ad dolores sit sit quod enim.
        end: "2017-08-22T12:06:30.696Z"
        id: 59b287534defd02dace3671b
        path: Quasi consequuntur.
        properties:
          Asperiores culpa quia.: Aperiam ea minima autem consequatur ducimus ut.
        refs:
        - Magni et facere sapiente voluptatem sunt quis.
        - Repudiandae voluptatem est.
        - Dolores consequatur ex repellendus.
        start: "2017-08-22T12:06:30.696Z"
        title: Veniam cumque aut.
      - deactivated: false
        description: Unde ad dolores sit sit quod enim.
        end: "2017-08-22T12:06:30.696Z"
        id: 59b287534defd02dace3671b
        path: Quasi consequuntur.
        properties:
          Asperiores culpa quia.: Aperiam ea minima autem consequatur ducimus ut.
        refs:
        - Magni et facere sapiente voluptatem sunt quis.
        - Repudiandae voluptatem est.
        - Dolores consequatur ex repellendus.
        start: "2017-08-22T12:06:30.696Z"
        title: Veniam cumque aut.
      - deactivated: false
        description: Unde ad dolores sit sit quod enim.
        end: "2017-08-22T12:06:30.696Z"
        id: 59b287534defd02dace3671b
        path: Quasi consequuntur.
        properties:
          Asperiores culpa quia.: Aperiam ea minima autem consequatur ducimus ut.
        refs:
        - Magni et facere sapiente voluptatem sunt quis.
        - Repudiandae voluptatem est.
        - Dolores consequatur ex repellendus.
        start: "2017-08-22T12:06:30.696Z"
        title: Veniam cumque aut.
      - deactivated: false
        description: Unde ad dolores sit sit quod enim.
        end: "2017-08-22T12:06:30.696Z"
        id: 59b287534defd02dace3671b
        path: Quasi consequuntur.
        properties:
          Asperiores culpa quia.: Aperiam ea minima autem consequatur ducimus ut.
        refs:
        - Magni et facere sapiente voluptatem sunt quis.
        - Repudiandae voluptatem est.
        - Dolores consequatur ex repellendus.
        start: "2017-08-22T12:06:30.696Z"
        title: Veniam cumque aut.
    CspaceEventResponseSummaryCollection:
      type: array
      items:
        $ref: '#/components/schemas/ShowResponseBodySummary2'
      description: TimeSearchResponseBody is the result type for an array of CspaceEventResponse
        (summary view)
      example:
      - end: "2017-08-22T12:06:30.696Z"
        id: 59b287534defd02dace3671b
        path: Voluptatem fugiat nemo facere cumque.
        start: "2017-08-22T12:06:30.696Z"
        title: Nam aspernatur enim molestiae laboriosam.
      - end: "2017-08-22T12:06:30.696Z"
        id: 59b287534defd02dace3671b
        path: Voluptatem fugiat nemo facere cumque.
        start: "2017-08-22T12:06:30.696Z"
        title: Nam aspernatur enim molestiae laboriosam.
...
    CspaceOrganismResponseSummaryCollection:
      type: array
      items:
        $ref: '#/components/schemas/ShowResponseBodySummary3'
      description: ListByPathResponseBody is the result type for an array of CspaceOrganismResponse
        (summary view)
      example:
      - id: 59b287534defd02dace3671b
        path: Porro atque neque adipisci.
        title: Dolores hic optio id.
      - id: 59b287534defd02dace3671b
        path: Porro atque neque adipisci.
        title: Dolores hic optio id.
...
    Error:
      type: object
      properties:
        fault:
          type: boolean
          description: Is the error a server-side fault?
          example: true
        id:
          type: string
          description: ID is a unique identifier for this particular occurrence of
            the problem.
          example: 123abc
        message:
          type: string
          description: Message is a human-readable explanation specific to this occurrence
            of the problem.
          example: parameter 'p' must be an integer
        name:
          type: string
          description: Name is the name of this class of errors.
          example: bad_request
        temporary:
          type: boolean
          description: Is the error temporary?
          example: true
        timeout:
          type: boolean
          description: Is the error a timeout?
          example: false
      example:
        id: 3F1FKVRR
        message: Value of ID must be an integer
        name: bad_request
      required:
      - name
      - id
      - message
      - temporary
      - timeout
      - fault
...